+++ changelog.md
@@ -2,6 +2,7 @@
 
 * Configurable fuel consumption can now exceed 1
 * If fuel consumption is set to 0, no initial item needs to get burned
+* Don't burn fuel in creative mode (configurable)
 
 # 0.4.1